// multiframework-casemanager/force-app/main/default/uiBundles/CaseManager/src/api/graphql-operations-types.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Base64: { input: string; output: string; }
  Currency: { input: number | string; output: number; }
  Date: { input: string; output: string; }
  DateTime: { input: string; output: string; }
  Double: { input: number | string; output: number; }
  Email: { input: string; output: string; }
  EncryptedString: { input: string; output: string; }
  /** Can be set to an ID or a Reference to the result of another mutation operation. */
  IdOrRef: { input: string; output: string; }
  Latitude: { input: number | string; output: number; }
  /** A 64-bit signed integer */
  Long: { input: number; output: number; }
  LongTextArea: { input: string; output: string; }
  Longitude: { input: number | string; output: number; }
  MultiPicklist: { input: string; output: string; }
  Percent: { input: number | string; output: number; }
  PhoneNumber: { input: string; output: string; }
  Picklist: { input: string; output: string; }
  RichTextArea: { input: string; output: string; }
  TextArea: { input: string; output: string; }
  Time: { input: string; output: string; }
  Url: { input: string; output: string; }
};
// ...
export type CreateCaseMutationVariables = Exact<{
  subject: Scalars['String']['input'];
  status: Scalars['Picklist']['input'];
  priority: Scalars['Picklist']['input'];
}>;


export type CreateCaseMutation = { uiapi: { CaseCreate?: { Record?: { Id: string, CaseNumber?: { value?: string | null } | null, Subject?: { value?: string | null } | null, Status?: { value?: string | null } | null, Priority?: { value?: string | null } | null, CreatedDate?: { value?: string | null } | null } | null } | null } };

export type DeleteCaseMutationVariables = Exact<{
  id: Scalars['IdOrRef']['input'];
}>;


export type DeleteCaseMutation = { uiapi: { CaseDelete?: { Id?: string | null } | null } };

export type UpdateCaseMutationVariables = Exact<{
  id: Scalars['IdOrRef']['input'];
  status: Scalars['Picklist']['input'];
}>;


export type UpdateCaseMutation = { uiapi: { CaseUpdate?: { success?: boolean | null, Record?: { Id: string, CaseNumber?: { value?: string | null } | null, Status?: { value?: string | null } | null } | null } | null } };

export type GetCasesQueryVariables = Exact<{ [key: string]: never; }>;


export type GetCasesQuery = { uiapi: { query: { Case?: { totalCount: number, edges?: Array<{ node?: { Id: string, CaseNumber?: { value?: string | null } | null, Subject?: { value?: string | null } | null, Status?: { value?: string | null } | null, Priority?: { value?: string | null } | null, CreatedDate?: { value?: string | null } | null } | null } | null> | null } | null } } };